[feat]: chatRoom 리스트 컴포넌트 추가
